requirements.txt
addict==2.4.0
asttokens==2.4.1
bidict==0.23.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
facenet-pytorch==2.5.3
filelock==3.13.4
fonttools==4.51.0
fsspec==2024.3.1
h11==0.14.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
intel-openmp==2021.4.0
ipykernel==6.29.4
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
lapx==0.5.9.post1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl==2021.4.0
mmcv==2.1.0
mmengine==0.10.4
mpmath==1.3.0
nest-asyncio==1.6.0
netifaces==0.10.6
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
py-cpuinfo==9.0.0
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-engineio==4.9.1
python-socketio==5.11.3
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.0
regex==2024.4.16
requests==2.31.0
rich==13.7.1
scipy==1.13.1
seaborn==0.13.2
simple-websocket==1.0.0
six==1.16.0
stack-data==0.6.3
sympy==1.12
tbb==2021.11.0
termcolor==2.4.0
tomli==2.0.1
torch==2.3.0+cu118
torchaudio==2.3.0+cu118
torchvision==0.18.0+cu118
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
typing_extensions==4.11.0
tzdata==2024.1
ultralytics==8.2.46
ultralytics-thop==2.0.0
urllib3==2.2.1
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
yapf==0.40.2
zipp==3.18.1
